requirements.txt
Django==2.2.10
django-rest-framework==0.1.0
djangorestframework==3.11.0
mock==4.0.2
pytz==2020.1
sqlparse==0.3.1
